requirements.txt
scipy==1.0.0
requests==2.13.0
pandas==0.22.0
Django==1.11.2
pymongo==3.6.1
djangorestframework_simplejwt==3.2.3
django-cors-headers==2.2.0
django-crispy-forms==1.6.1
django-filter==1.0.4
django-filters==0.2.1
django-heroku==0.3.1
django-markdown-deux==1.0.5
django-pagedown==0.1.3
djangorestframework-jwt==1.10.0
numpy==1.13.3
djangorestframework==3.8.2
scikit_learn==0.19.1
